Got the controller logic setup and an example of how to pass things to create

+++ db/migrate/20140819043007_create_transactions.rb
@@ -2,7 +2,7 @@ class CreateTransactions < ActiveRecord::Migration
   def change
     create_table :transactions do |t|
       t.belongs_to :company, null: false
-      t.long :amount, null: false
+      t.integer :amount, null: false
       t.string :charge_id, null: false
       t.text :description
